requirements.txt
asgiref==3.7.2
blinker==1.6.2
channels==4.0.0
click==8.1.7
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.5
django-plotly-dash==2.2.2
dpd-components==0.1.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
paho-mqtt==1.6.1
plotly==5.17.0
sqlparse==0.4.4
tenacity==8.2.3
Werkzeug==2.3.7
